Work
Added Cash but not betting and help me Timmy is close and I believe he is going to harm me, he detects bias in Mr.Morraye, his burner is very complex but Mr.Morraye will not simply allow him to get one. He says Python is open source now he is ranting and starting to be an angery boy.